Added more code modularity and updated typescript use
