+++ src/App.tsx
@@ -13,63 +13,21 @@ import CreateRecipe from "./components/CreateRecipe";
 import Login from "./components/Login";
 
 // Theme
-import { createTheme, ThemeProvider } from "@mui/material/styles";
+import theme from "./AppTheme";
+import { ThemeProvider } from "@mui/material/styles";
 
 // Material Ui
 import { Grid } from "@mui/material";
 import { Auth } from 'aws-amplify';
 
 //Models
-import { RecipeI } from "./model";
+import { RecipeInterface } from "./model";
 
-const theme = createTheme({
-  palette: {
-    mode: "dark",
-    background: {
-      default: "#2d2d37",
-      paper: "#212226",
-    },
-    text: {
-      primary: "#ffffff",
-    },
-    primary: {
-      main: "#000",
-    },
-    secondary: {
-      main: "#fff",
-    },
-  },
-  typography: {
-    allVariants: {
-      color: "white",
-    },
-  },
-  components: {
-    MuiAppBar: {
-      styleOverrides: {
-        colorPrimary: {
-          backgroundColor: "#000",
-          color: "#000",
-        },
-      },
-    },
-    MuiInputLabel: {
-      styleOverrides: {
-        root: {
-          fontWeight: "bold",
-          "&.Mui-focused": {
-            color: "white",
-          },
-        },
-      },
-    },
-  },
-});
 
 const App: React.FC = () => {
   
   const [userData, setUserData] = useState<any>(null);
-  const [recipes, setRecipes] = useState<RecipeI[] | null>(null);
+  const [recipes, setRecipes] = useState<RecipeInterface[] | null>(null);
 
   const checkForUser = async () => {
     try {
